Added tests for new device properties
Reflactor common duplicated fixtures for device properties checks
out into a common file made available to both `test_sycl_device.py`
and `test_sycl_queue.py` via `conftest.py` per

https://stackoverflow.com/questions/13641973/how-and-where-does-py-test-find-fixtures

MaxWorkItemSizes1d, MaxWorkItemSizes2d and MaxWorkItemSizes3d properties
are added.
